Refactored existing code into basis for a module system
This includes reworking how the code is organized so that we can maintain separation of all the pieces. I also pulled in a (relatively) default .jshintrc file so we have ES6 hinting in place.

NOTE that we can't use proper ES6 modules in Chrome stable just yet, so I've left that out.