Add ability to filter gradebook data
I changed how gradebook data was collected. It now includes an assignment's due date. According to Tiyo, an assignment is only required if it has a due date and it's not hidden.

I changed how the gradebook is rendered. Assignments are now listed left to right, with the oldest shown to the left. (This could be reversed, obviously.) I added checkboxes that allow instructors to select what assignments to display. Combinations of selections can be made. For example, if I check all four checkboxes all assignments show, even those that I have no intention of giving to students.

As assignments are shown or hidden student grades are updated. This allows me to uncheck all boxes to see only the assignments that are open and have submissions and are required. I can easily tell what students are doing well and which are not, without extraneous information. Additionally, I can see how far a given student is in the entire class.

I changed the display styles a bit too. For example, there is now a vertical line separating grades. The column header is centered directly above the column it relates to.

Lastly, I updated the display so that if I select the entire table and copy and paste it into a spreadsheet that I can see the full name of the assignment, not just the first letter.

One last note, the indenting in the gradebook.js was inconsistent. In some places it used two spaces, in others it used tabs. I tried to follow the conventions used in any particular section, as much as possible, without reformatting the entire file.