Update gradebook styles
I changed the display styles for gradebooks. There is now a vertical line separating grades. The column header is centered directly above the column it relates to.

Additionally, if the gradebook table becomes too wide to fit on the page it will scroll. This is accomplished by wrapping the table in a div with an overflow-x of scroll. This change caused some problems with how tooltips are displayed in the column headers. To resolve this problem I added 1 em of padding in the div and -1em margin top. All in all, the tooltips look the same as before.

Column labels are centered horizontally now too.