Update column headers to include hidden full assignment name
This commit updates the headers of the columns in the gradebook to contain the full text of each assignment's name. This is styled such that only the first letter is visible and all of the remaining letters are hidden.

This has utility in that copying and pasting the content of the table to a spreadsheet will now include the full title of the assignment.